Start to setup image library
