+++ libraries/gpu/src/gpu/Format.h
@@ -165,12 +165,11 @@ enum Semantic {
     COMPRESSED_SRGB,
     COMPRESSED_SRGBA,
 
-    // FIXME: Will have to be supported later:
-    /*COMPRESSED_BC3_RGBA,  // RGBA_S3TC_DXT5_EXT,
+    COMPRESSED_BC3_RGBA,  // RGBA_S3TC_DXT5_EXT,
     COMPRESSED_BC3_SRGBA, // SRGB_ALPHA_S3TC_DXT5_EXT
 
     COMPRESSED_BC7_RGBA,
-    COMPRESSED_BC7_SRGBA, */
+    COMPRESSED_BC7_SRGBA,
 
     _LAST_COMPRESSED,
