+++ libraries/image/CMakeLists.txt
@@ -1,0 +1,9 @@
+set(TARGET_NAME image)
+setup_hifi_library()
+link_hifi_libraries(shared gpu)
+
+target_glm()
+
+add_dependency_external_projects(nvtt)
+target_include_directories(${TARGET_NAME} PRIVATE ${NVTT_INCLUDE_DIRS})
+target_link_libraries(${TARGET_NAME} ${NVTT_LIBRARIES})
