+++ libraries/image/src/image/ImageLogging.cpp
@@ -1,0 +1,14 @@
+//
+//  ImageLogging.cpp
+//  image/src/image
+//
+//  Created by Clement Brisset on 4/5/2017.
+//  Copyright 2017 High Fidelity, Inc.
+//
+//  Distributed under the Apache License, Version 2.0.
+//  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
+//
+
+#include "ImageLogging.h"
+
+Q_LOGGING_CATEGORY(imagelogging, "hifi.model")
