+++ libraries/model-networking/CMakeLists.txt
@@ -1,4 +1,4 @@
 set(TARGET_NAME model-networking)
 setup_hifi_library()
-link_hifi_libraries(shared networking model fbx ktx)
+link_hifi_libraries(shared networking model fbx ktx image)
