+++ libraries/model/src/model/TextureMap.h
@@ -13,47 +13,9 @@
 
 #include "gpu/Texture.h"
 
-#include "Material.h"
 #include "Transform.h"
 
-#include <qurl.h>
-
-class QImage;
-
 namespace model {
-
-typedef glm::vec3 Color;
-
-class TextureUsage {
-public:
-    gpu::Texture::Type _type{ gpu::Texture::TEX_2D };
-    Material::MapFlags _materialUsage{ MaterialKey::ALBEDO_MAP };
-
-    int _environmentUsage = 0;
-
-    static gpu::Texture* create2DTextureFromImage(const QImage& image, const std::string& srcImageName);
-    static gpu::Texture* createStrict2DTextureFromImage(const QImage& image, const std::string& srcImageName);
-    static gpu::Texture* createAlbedoTextureFromImage(const QImage& image, const std::string& srcImageName);
-    static gpu::Texture* createEmissiveTextureFromImage(const QImage& image, const std::string& srcImageName);
-    static gpu::Texture* createNormalTextureFromNormalImage(const QImage& image, const std::string& srcImageName);
-    static gpu::Texture* createNormalTextureFromBumpImage(const QImage& image, const std::string& srcImageName);
-    static gpu::Texture* createRoughnessTextureFromImage(const QImage& image, const std::string& srcImageName);
-    static gpu::Texture* createRoughnessTextureFromGlossImage(const QImage& image, const std::string& srcImageName);
-    static gpu::Texture* createMetallicTextureFromImage(const QImage& image, const std::string& srcImageName);
-    static gpu::Texture* createCubeTextureFromImage(const QImage& image, const std::string& srcImageName);
-    static gpu::Texture* createCubeTextureFromImageWithoutIrradiance(const QImage& image, const std::string& srcImageName);
-    static gpu::Texture* createLightmapTextureFromImage(const QImage& image, const std::string& srcImageName);
-
-
-    static const QImage process2DImageColor(const QImage& srcImage, bool& validAlpha, bool& alphaAsMask);
-    static void defineColorTexelFormats(gpu::Element& formatGPU, gpu::Element& formatMip,
-        const QImage& srcImage, bool isLinear, bool doCompress);
-    static gpu::Texture* process2DTextureColorFromImage(const QImage& srcImage, const std::string& srcImageName, bool isLinear, bool doCompress, bool generateMips, bool isStrict = false);
-    static gpu::Texture* processCubeTextureColorFromImage(const QImage& srcImage, const std::string& srcImageName, bool isLinear, bool doCompress, bool generateMips, bool generateIrradiance);
-
-};
-
-
 
 class TextureMap {
 public:
